Make statement handle storage threadsafe
Connection objects retain a map of ongoing statements. With SQLite,
one can use a single database handle to execute queries from many
threads when it is in serialized mode (which is the default). This
works out OK with SQLite itself, however the handling of the map of
active statement handles was not threadsafe, and this could cause
segfaults.

This refactors how active statement handles are tracked to allow the
insertion of locking, and then adds that locking.