Temporary failure error handling
DBs have numerous temporary errors from deadlocks to serialization failures to network failures. SQLState contains the error code (largely standardized across SQL DBs, but not completely) for various situations which may require a retry or allow the client to perform other correctable work.

Export sqlstate and all other available fields for Pg in X::DBDish::DBError::Pg (extends X::DBDish::DBError).

While here, shorten the message to just the error component for Pg. Detail and Hint are available via other fields and is quite noisy (it typically starts with Error: ERROR: as DBIish adds an 'Error:' prefix too).

Add is-temporary which checks if the sqlstate for Pg is a temporary failure and the transaction should be retried by the client. These states include:
  Class 08\w{3}: All connection exceptions (possible network issues)
  40001 serialization_failure
  40P01 deadlock_detected
  Class 57\w{3}: Operator Intervention (early/forced connection termination)
  72000 snapshot_too_old

The intent is that a tool like Cro can check for is-temporary across all DB drivers and restart working on the request with very little assistance from the developer other than perhaps a "sub routes(:restartable)" type markup to indicate their code is safe to re-execute.