Quieten non-problematic worry in t/36-pg-enum.t
We were getting:

	Potential difficulties:
	    No values supplied to enum (does @options need to be declared constant?)
	    at /home/jonathan/devel/perl6/3rdparty-modules/DBIish/t/36-pg-enum.t:74
	    ------> my enum YesNo (@options)⏏;

because the @options that is passed to the enum is empty at compile time
and is populated at run-time from values from the DB.  The test was
otherwise fine.

Developers may want to temporarily comment out the "no worries" while
working on this file or the Pg driver as it will suppress worries about
new problems.

Fixes #116