Use much cheaper conditional method call construct
Setting the last error - or more typically clearing the last error - is
an extremely common operation. The `.^can` meta-call returns a list of
methods that match, not just a boolean, and so takes quite a bit of time
if done repeatedly. However, the same effect can be achieved with the
`.?` conditional method call operator, which is well optimized. In a
benchmark doing a hundred thousand executes of the same prepared insert,
this gets the benchmark to run in 46% of the time it previously did.