Avoid repeated use of mixins
A mixin causes a global deoptimization at present, which means its costs
are quite high and non-local. We only really need to provide a true 0
when we are actually returning 0, and we can also produce that at
compile time, meaning there's no runtime work at all at runtime. This
gets a hundred thousand insert execute benchmark running in 78% of the
time it did before this change.