Protect from loading drivers multiple times
Multiple threads establishing a database connection for the first time in the script simultaneously will stomp on each-other while modifying %installed{$drivername}. Prevent this with a lock so the driver is only loaded once.

The sqlite threading example doesn't trigger this issue as the connection is tested prior to multiple threads being started. Create a Pg test which does module loading after the threads have been started.