Lock connection while in use
A generic lock/unlock (via atomics) has been provided for all database types (this is an issue on MySQL, SQLite, and Oracle too) but there is no generic way to implement it on everything. The success case would be straight forward in DBDish::StatementHandle but error cases make unlock a bit trickier.

Implement on PostgreSQL to protect against corrupting the database connection.

As a future addition, there might be a generic ability to register a promise for improved multi-threaded interactions with a single DB connection.