Enable caller to do DB work within CATCH block
Corrects an oversight in c1d3c7c.

This is a fairly common pattern:

   try {
     $dbh.do('BEGIN');
     $dbh.do('INSERT...');
     $dbh.do('INSERT...');
     CATCH {
       $dbh.do('ROLLBACK');
     }
   }

In the previous coding the CATCH for the caller above would execute prior to the LEAVE in protect-connection. This caused the ROLLBACK to issue a concurrency error where there wasn't one. Tighten protect-connection by unlocking in a CATCH block and in the success case.