Protect connection from multi-threaded use
During a protocol corruption the database will disconnect. Locking the connection protects against protocol corruption.

This locking may not be sufficient when mysql_use_result() is used (Prefetch => False), however this is not the default mode for statement handles so leave that for a different patch. The other, perhaps better, option is to remove the Prefetch option and always use the safer mysql_store_result().