PostgreSQL interval is not a Duration
An interval isn't necessarily just seconds. It can be something more complicated like this:

  select interval '1 year 1 month -5 days 10 hours 5 seconds -10 milliseconds';
  => 1 year 1 mon -5 days +10:00:04.99

Also, to parse the driver would need to force the connection to be in a specific interval format, as there are several:

  set intervalstyle = 'iso_8601';
  select interval '1 year 1 month -5 days 10 hours 5 seconds -10 milliseconds';
  => P1Y1M-5DT10H4.99S

So a naive number only conversion fails outright.

The iso_8601 style above may seem straight forward to parse you still need to figure out how many seconds are in a year, month, and day; that requires a reference date as a starting point including the time zone.

The conversion currently fails on large intervals; and would be an approximation at best if we attempted a conversion. A user willing to take an approximate conversion can use extract(EPOCH <interval>) on the PostgreSQL side.