Fix parsing seemingly mixed data in arrays
Values like mac addresses were matching as integers for the first portion then as string for the rest.

This SQL array with a single element:
  ARRAY['12:34:56:78']

Was being interpreted by Raku as:
   ['12', ':34:56:78']

To fix, separate handling of quoted strings from unquoted strings. Unquoted strings are tested for type as a separate step after the entire element is extracted.