Deprecate quote-identifier; prefer quote(:as-id)
The MySQL test stated that this was the intended direction and the Pg driver was implemented that way. Create a generic quote() which implemented SQL spec escaping (doubles the ' or " as needed). Drivers should use routines provided by the database if they are available for safety.

Document quote and show an example of how to use it.

Closes #193