Use the converter for simple query results
When I pushed 96a0e4d I left this as a step for the future, mistakenly believing it was for complex user provided conversions only. Basics like DateTime also require the conversion functionality.

The original type conversion for stringy data in want() was a bit too simple. Replacing it in-location with functions that users may provide also felt like the wrong place. A refactoring of _row() and MyRow may be desirable but leave that for the future as fetch_row result-sets may require additional work.

Closes #196