Improve performance of large results
By handling the type selection once per query instead of once per record, quite a bit of time is shaved off the conversion in MoarVM 2021.10.

For this sample bit of code, execution time goes from ~80 seconds to ~40 seconds. While a good improvement, similar code in Perl runs in a sub-second time using DBD::Pg.

  my constant $records = 1_000_000;

  my $sth = $dbh.execute(q{SELECT col, col as col2, col as col3 FROM generate_series(1, $1) AS k(col)}, $records);

  my $sum = 0;
  while (my $row = $sth.row()) {
    $sum += $row[0] + $row[1] - $row[2];
  }

After this commit List::sink and QAST::decont_all take 75% of the runtime. This may be an issue with the profiler rather than real runtime.