Handle type selection once for MySQL driver
This is the same technique as the PostgreSQL driver in c656ee5. Setup type conversion functions once per result set instead of hunting for them once per record.

Iterating on @!import-func and incrementing $col seems to perform the best, though it's marginal due to overhead in the rest of the driver.