Allow changing column-type manually again
There is an undocumented and very lightly tested feature which allows the column-type to be specified by the client after the prepare/execute phase of the query but before retrieving rows. Re-enable this again for the PostgreSQL driver by setting up import functions on the first (and only the first) call to row().