Quickfix for xt/ tests
Relied upon diag-functions routine used to be provided by
NativeCall::TypeDiag but which is long not there.

Needed for `mi6 release` which tries to run tests in xt/ too.

NOTE. It looks like nobody did xt/ testing in years since the last
commit to NativeCall::TypeDiag is dated back to 2016! Perhaps it makes
more sense to remove them altogether?