Use mysql_exec_statement API whenever possible
The list of statements supported by the prepared statement API includes most things:
https://dev.mysql.com/doc/refman/8.0/en/sql-prepared-statements.html#prepared-statements-permitted

Consistently using this API when possible provides consistency with NULL handling and potentially other divergences which we've not yet discovered.

A few, such as LOCK TABLE or LOAD TABLE or SHOW WARNINGS, are yet supported so fallback to the older query mechanism when required. If performance with this type of query is an issue, peaking at the first word might be possible rather than relying strictly on the try/catch.